% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{build_manifest}
\alias{build_manifest}
\title{Build file manifest from 10X cellranger-atac}
\usage{
build_manifest(pattern = NULL)
}
\arguments{
\item{pattern}{Common string among all samples}
}
\description{
Assumes that each sample is in its own folder. Pattern match to get multiple samples
Returns named list where each sample has paths to peaks.bed, singlecell.csv, and fragments file
Also possible to build list or make corrections manually
}
\examples{
m = build_manifest(pattern = "18655")

}
