% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{export_umap}
\alias{export_umap}
\title{Export UMAP embeddings from a Suerat object}
\usage{
export_umap(
  srt = srt,
  embeddings_file = "umap_embeddings.csv",
  cluster_id_file = "clusters.csv",
  Idents = NULL,
  drop_cell_ids = TRUE
)
}
\arguments{
\item{srt}{A Seurat object}

\item{cluster_id_file}{Name of csv file that containst he cluster IDs}

\item{Idents}{Set the active Ident from Seurat metadata before exporting. Becomes grouping factor in clusters.csv. Default is active Ident.}

\item{drop_cell_ids}{Drop the cell ID prefixed the the cell barcode by Seurat}

\item{umap_embeddings_file}{Name of csv file that contains UMAP embeddings}
}
\description{
Extract UMAP embedding from a Suerat object and write to file as csv
By defult, writes two files, umap_embeddings.csv and clusters.csv for use in Loupe browser
By default, drops a cell ID that is prefixed the cell barcode
}
\examples{
\dontrun{
export_umap(srt = srt, embeddings_file = "umap_embeddings.csv", cluster_id_file = "clusters.csv",
    Idents = NULL, drop_cell_ids = TRUE)
}

}
