% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_to_human.R
\name{loupe_mouse_to_human}
\alias{loupe_mouse_to_human}
\title{Convert Loupe differential gene expression results from mouse to human homologs for GSEA}
\usage{
loupe_mouse_to_human(
  loupe_export = NULL,
  mouse_db = hciRdata::mouse102,
  only_named = TRUE
)
}
\arguments{
\item{loupe_export}{The CSV table exported from Loupe}

\item{mouse_db}{The mouse database with human homologs. Default is hciRdata::mouse102}

\item{only_named}{Return only genes with human homologs. Defalt is TRUE}
}
\description{
After running differential gene expression in Loupe with mouse data,
gene names need to be converted to human homologs for GSEA. Returns
list where each element is a cluster from the Loupe export
}
