% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{percent_table}
\alias{percent_table}
\title{Convert a counts table to fraction of total observations}
\usage{
percent_table(df = df, by = "Row", transpose = TRUE)
}
\arguments{
\item{df}{Either a data frame or a table produced by base::table()}

\item{by}{Input vectors are either "Row" or "Column"}

\item{transpose}{Transpose resulting table back to original orientation}
}
\description{
Use either rowSums or colSums to convert each cell to a fraction of total observations by row or columns
}
\examples{
\dontrun{
t = percent_table(df = df, by = "Row", transpose = TRUE)
}

}
