% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{prep_sc_DE_for_fgsea}
\alias{prep_sc_DE_for_fgsea}
\title{Covert the ouptut of Seurat's FindMarkers to a table that hciR's fgsea_all can use}
\usage{
prep_sc_DE_for_fgsea(
  markers = markers,
  contrast = NULL,
  database = hciRdata::human100,
  human_homolog = FALSE,
  protein_coding = TRUE
)
}
\arguments{
\item{markers}{Table from Seurat's FindMarkers}

\item{contrast}{The name of the contrast}

\item{database}{Table of genes to use when gathering annotations. Default is hciRdata::human100}

\item{human_homolog}{Convert the gene names to human homologs. Default is FALSE}
}
\description{
This function assumes that a user has run Seurat's FindMarkers to
assess differential expression between clusters and would like to
next run GSEA with hciR's fgsea_all and gene to pathway data from
hciRdata.
}
\examples{
\dontrun{
sc_results = prep_sc_DE_for_fgsea(markers = markers, contrast = "cluster 4 vs. cluster 0")
fgsea_all(res = sc_results, gsets = msig_pathways$KEGG, FDR = 0.1, nperm = 10000)
}

}
