% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read10X_multiple}
\alias{read10X_multiple}
\title{Read in multiple 10X scRNAsea samples}
\usage{
read10X_multiple(pattern = NULL, min_cells = 0, min_features = 0, outs = FALSE)
}
\arguments{
\item{pattern}{The pattern to search for in current working dicrectory, e.g, run number}

\item{min_features}{Seurat min.features}

\item{outs}{Include the "outs" in the full path to filtered barcode files. Default is FALSE}

\item{min_cell}{Seurat min.cells}
}
\description{
Searches for pattern in current working directory and returns list of Seurat objects
}
